fix: make self-derived key derivation match older addresses
This modifies generation_address so that the same addresses will
be derived from a given wallet secret as happened at tag alphanet-v5.

A test exists which verifies this:
  wallet_tests::generation_key_derivation::verify_derived_generation_addrs()

Significant changes:

* compatibility with alphanet-v5 addresses:
    * DerivationIndex: u128 --> u64
    * replace GenerationSpendingKey::seed with secret, index
    * rename GenerationSpendingKey::from_seed() --> from_secret_and_index()
    * add GenerationSpendingKey::seed(secret, index) which uses same
       hash(secret, generation_flag, index) construction as
       WalletSecret::nth_generation_key() used to.
    * WalletSecret::master_generation_key is now DerivationIndex::MAX so
      that the 0th key can have a parent.
    * GenerationSpendingKey::derive_child() now uses wrapping_add() so that
      master key (index = DerivationIndex::MAX) can derive 0th key.

* add KeyTypeSeed to abstract over seed types

Minor changes:
* update a couple Block tests to use mainnet.
* rename SpendingKeyIter::curr      --> curr_fwd
* rename SpendingKeyIter::curr_back --> curr_rev
* updates test verify_derived_generation_keys with newly generated keys
  because serialized key format has changed.
* add some missing doc-comments for key/addr in generation_address.rs
* update GenerationSpendingKey deserialization impl
* add impl From<&GenerationSpendingKey> for GenerationReceivingAddress
* remove GenerationReceivingAddress::from_spending_key()
* update Arbitrary impls for GenerationSpendingKey seeds
* add a couple of Arbitrary impl to the feature flag.
* adapt tests to changes

+++ src/models/state/wallet/address/common.rs
@@ -13,7 +13,7 @@ use crate::models::blockchain::transaction::PublicAnnouncement;
 use crate::models::state::wallet::utxo_notification::UtxoNotificationPayload;
 use crate::prelude::twenty_first;
 
-pub type DerivationIndex = u128;
+pub type DerivationIndex = u64;
 
 /// returns human-readable-prefix for the given network
 pub(crate) fn network_hrp_char(network: Network) -> char {
